Move annotation mixin in the mixin folder
It was kind of lost in the preview folder
