+++ src/components/mixins/annotation_mixin.js
